Change project.function purpose
It seems like the whole script for a project is currently being
held in the function field of the project table, but really
all that is needed is to copy it from the corresponding
script file. A redirect or even going to the right file directly
would also work better.